soc and other misc changes
