+++ core/static/css/post.css
@@ -10,7 +10,6 @@ main {
     align-self: center;
     width: 90%;
     max-width: 800px;
-    /* margin: 2rem auto; */
     padding: 0 1rem;
 }
