+++ core/templates/core/blogpost.html
@@ -5,9 +5,11 @@
     <meta name='robots' content='noindex,follow' />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.1/jquery.min.js" integrity="sha512-v2CJ7UaYy4JwqLDIrZUI/4hqeoQieOmAZNXBeQyjo21dadnwR+8ZaIJVT8EE2iyI61OV8e6M8PP2/4hpQINQ/g==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
+    <script src="{% static 'js/highlight.min.js' %}"></script>
     <link rel="stylesheet" href="{% static 'css/base.css' %}">
     <link rel="stylesheet" href="{% static 'css/post.css' %}">
     <link rel="stylesheet" href="{% static 'css/table-of-contents.css' %}">
+    <link rel="stylesheet" href="{% static 'css/highlight-css/github-dark-dimmed.min.css' %}">
 </head>
 <body>
     {% include 'core/components/header.html' %}
@@ -29,7 +31,6 @@
                 {{ post.body|safe}}
             </div>
         </main>
-    <!-- <p>Related Posts</p> -->
     <div class="related-posts">
         {% for post in related_posts %}
         <div class="post">
@@ -47,4 +48,5 @@
         {% endfor %}
     </div>
     {% include 'core/components/footer.html' %}
+    <script>hljs.highlightAll();</script>
 </body>
