--- core/templates/core/components/dump.html
@@ -1,21 +1,0 @@
-// .then(data => {
-    //     if(Array.isArray(data)) {
-    //         data.forEach(ele => {
-    //             post = `<div class="post">
-    //                         <h2><a href="${window.location.hostname}/posts/${ele.slug}/">${ele.title}</a>
-    //                         <div class="post_detail">
-    //                             ${ele.date_created_on}</br>
-    //                             <div class="post_categories">
-    //                             ${ele.categories.forEach(category => {
-    //                                 `<a href="${window.location.hostname}/category/${category}" data-anchor="category">${category}</a>`
-    //                             })}
-    //                             </div>
-    //                         </div>
-    //                     </div>`
-                
-    //             post_container.innerHTML += post;   
-    //         });
-    //     }
-    //     else {
-    //         console.error("data is not array", data);
-    //     }
