+++ core/views.py
@@ -1,17 +1,13 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import BlogPost, Category
+from .models import BlogPost
 from .forms import BlogPostForm
-from django.contrib.auth.models import User
 from django.http import HttpResponseForbidden, JsonResponse
 from django.template.loader import render_to_string
 from django.db.models import Q
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-
 from functools import wraps
-from .utils import is_mobile_device, is_ajax, fetchQuote
-
+from .utils import is_mobile_device, is_ajax, fetchQuote, BlogDetail, BlogPostService
 import json
-# Create your views here.
+
 def superuser(view):
     @wraps(view)
     def _wrapped_view(request, *args, **kwargs):
@@ -47,99 +43,10 @@ def fetch(request):
         }
         return render(request, 'core/terminal.html', context)
 
-class BlogPostService:
-    @staticmethod
-    def getPosts(user: User, pageNo: int, number_of_posts: int) -> list:
-        if user.is_superuser:
-            posts = BlogPost.objects.all()
-        else:
-            posts = BlogPost.objects.all().filter(password_protect=False)
-
-        paginatorInstance = Paginator(posts, number_of_posts)
-    
-        try:
-            postsPage = paginatorInstance.get_page(pageNo)
-        except PageNotAnInteger:
-            postsPage = paginatorInstance.get_page(1)
-        except EmptyPage:
-            postsPage = paginatorInstance.get_page(paginatorInstance.num_pages)
-            postsPage.adjusted_elided_pages = paginatorInstance.get_elided_page_range(pageNo)
-        return postsPage, paginatorInstance
-    
-    @staticmethod
-    def getPostsByCategory(user: User, category: Category, pageNo) -> list:
-        if user.is_superuser:
-            posts = BlogPost.objects.filter(Q(categories__name__icontains=category))
-        else:
-            posts = BlogPost.objects.filter(Q(categories__name__icontains=category) & Q(password_protect=False))
-        
-        paginatorInstance = Paginator(posts, 4)
-        
-        try:
-            postsPage = paginatorInstance.get_page(pageNo)
-        except PageNotAnInteger:
-            postsPage = paginatorInstance.get_page(1)
-        except EmptyPage:
-            postsPage = paginatorInstance.get_page(paginatorInstance.num_pages)
-            postsPage.adjusted_elided_pages = paginatorInstance.get_elided_page_range(pageNo)
-        return postsPage
-    
-    @staticmethod
-    def getRelatedPosts(user: User, slug) -> list:
-        post = get_object_or_404(BlogPost, slug=slug)
-        categories = post.categories.all()
-        
-        for category in categories:
-            related_posts = BlogPost.objects.filter(categories=category).exclude(slug=slug)
-        
-        if user.is_superuser is False:
-            related_posts.filter(password_protect=False)
-
-        return related_posts
-
-class BlogDetail:
-    @staticmethod
-    def getBlogDetailAdminView(request, slug):
-        post = get_object_or_404(BlogPost, slug=slug)
-        if post.password_protect:
-            if request.user.is_superuser:
-                pass
-            else:
-                return error403(request, exception=HttpResponseForbidden)
-        response = {
-            'title': post.title,
-            'share_link': post.get_share_url(),
-            'categories': post.categories.all(),
-            'created_on': post.created_on,
-            'last_modified': post.last_modified,
-            'body': post.body,
-            'slug': post.slug,
-            'req_toc': post.require_table_of_contents
-        }
-        return response
-    
-    @staticmethod
-    def getBlogDetailShareView(uuid):
-        post = get_object_or_404(BlogPost, share_token=uuid)
-        
-        response = {
-            'title': post.title,
-            'categories': post.categories.all(),
-            'created_on': post.created_on.date,
-            'last_modified': post.last_modified,
-            'body': post.body,
-            'req_toc': post.require_table_of_contents
-        }
-
-        return response
-
 def index(request):    
     pageNo = request.GET.get('page')
-
     is_mobile = is_mobile_device(request)
-
     posts, paginator = BlogPostService.getPosts(request.user, pageNo, 5 if is_mobile else 6)
-
     quote = fetchQuote()
 
     context = {
@@ -148,9 +55,7 @@ def index(request):
         'quote': quote,
     }
 
-    ajax = is_ajax(request)
-
-    if ajax:
+    if is_ajax(request):
         posts = render_to_string('core/components/partial_posts.html', context)
         pagination = render_to_string('core/components/pagination.html', context)
 
@@ -175,9 +80,7 @@ def viewBlogByCategory(request, category):
 
 def viewBlog(request, slug):
     post = BlogDetail.getBlogDetailAdminView(request, slug)
-
     related_posts = BlogPostService.getRelatedPosts(request.user, slug)
-
     quote = fetchQuote()
 
     context = {
@@ -226,9 +129,7 @@ def delete_post(slug):
 
 def share(request, uuid):
     quote = fetchQuote()
-    
     slug = BlogPost.objects.get(share_token=uuid).slug
-
     related_posts = BlogPostService.getRelatedPosts(request.user, slug)
 
     context = {
@@ -263,11 +164,7 @@ def search(request):
 
     quote = fetchQuote()
 
-    context = {
-        'quote': quote
-    }
-
-    return render(request, 'core/components/search.html', context)
+    return render(request, 'core/components/search.html', {'quote': quote})
 
 def error403(request, exception):
     return render(request, 'core/error/403.html')
