+++ requirements.txt
@@ -1,4 +1,4 @@
 asgiref==3.8.1
-Django==5.1.5
+Django>=5.1.8
 django-js-asset==3.0.1
 sqlparse==0.5.3
